almost final version, TODO change error messages, add comments and documentation

+++ src/tickets-trading.Application/ServiceHandlers/MoneyAddingHandler.cs
@@ -1,12 +1,10 @@
-using tickets_trading.Application.DatabaseAPI;
 using tickets_trading.Domain;
-using Microsoft.EntityFrameworkCore;
 
 namespace tickets_trading.Application.ServiceHandlers;
 
 public class MoneyAddingHandler(RegularUser user)
 {
-    public void Handle(int moneyToAdd)
+    public void Handle(long moneyToAdd)
     {
         user.AddMoney(moneyToAdd);
     }
